:art: One cast spelling in kernel scope, and delete the no-op casts
The tree carried 329 ``int(wp.tid())`` alongside 640 ``wp.int32(...)``, and 46
sites in 41 kernels across 22 modules cast a thread index with one spelling and
re-cast the same local with the other a few lines later.

Measured on Warp 1.16 before changing anything:

- ``int(x)`` and ``wp.int32(x)`` generate identical code. ``int`` compiles only
  because Warp writes an unconditional ``#define int(x) cast_int(x)`` into every
  generated module header, ``wp::int(x)`` not being valid C++.
- ``float(...)`` is NOT equivalent: inside a ``wp.Float``-generic @wp.func it is
  a hard compile error (``Input types must be the same, got ['float64',
  'float32']``), so every bare one silently forecloses genericising its
  function -- against section 14's "prefer dtype-generic @wp.func".
  ``type(x)(...)`` is the generic-safe spelling.
- No cast is load-bearing. A bare ``wp.tid()`` passes unchanged to a ``wp.int32``
  parameter, to a ``wp.Scalar`` generic and into a kernel-scope slice.

So: 362 ``int(...)`` promoted to ``wp.int32(...)``, 184 casts to the type the
value already had deleted (a ``wp.int32`` argument, a ``wp.array[wp.int32]``
element, a ``.shape[0]``, a tid local), and the 22 ``float(...)`` respelled
``wp.float32(...)`` -- each checked individually, none of them generic. A cast
of a bare *literal* is never redundant and none was touched: ``wp.int32(0)`` is
a mutable dynamic variable where ``0`` is a compile-time constant that freezes
the enclosing loop.

The tid cast stays, as the one declarative cast: it names the type of the index
the whole kernel is written against.

Also spells integer division ``//`` at 18 sites. On ``wp.int32`` the two
operators are the same operation and both truncate toward zero: measured
``[-8, -7, -1, 0, 1, 7, 8] / 3`` gives ``[-2, -2, 0, 0, 0, 2, 2]`` for both,
where CPython's ``//`` gives ``[-3, -3, -1, 0, 0, 2, 2]``. Consistent with the
``%`` rule section 5 already records. No live bug -- every dividend here is a
non-negative index -- but ``/`` on two int32s reads as real division, and the
CPython divergence is a porting hazard worth writing down.

Pins both with api_conventions check 16 (a bare int()/float() in a kernel or
@wp.func body), which also caught that the module docstring was still missing
check 15.

Gates: 2293 passed / 1 skipped, ruff clean, basedpyright 0 errors, 348 parity
pairs with 0 uncovered, mkdocs --strict builds.

+++ triwarp/kernels/algorithms/blue_noise.py
@@ -93,7 +93,7 @@ def init_point_cells(
     out_point_cell: wp.array[wp.int32],
 ) -> None:
     # Compacted cell index of every pool point (its cell is occupied by construction).
-    i = int(wp.tid())
+    i = wp.int32(wp.tid())
     out_point_cell[i] = lookup_cell(unique_keys, grid_cell_key(grid_coords[i], grid_w))
 
 
@@ -110,9 +110,9 @@ def dart_cell_neighbors(
     x = wp.int32(key % w64)
     y = wp.int32((key / w64) % w64)
     z = wp.int32(key / (w64 * w64))
-    dx = wp.int32(s) % _DART_SHELL_W - wp.int32(1)
-    dy = (wp.int32(s) / _DART_SHELL_W) % _DART_SHELL_W - wp.int32(1)
-    dz = wp.int32(s) / (_DART_SHELL_W * _DART_SHELL_W) - wp.int32(1)
+    dx = s % _DART_SHELL_W - wp.int32(1)
+    dy = (s / _DART_SHELL_W) % _DART_SHELL_W - wp.int32(1)
+    dz = s / (_DART_SHELL_W * _DART_SHELL_W) - wp.int32(1)
     out_cell_neighbors[c, s] = INVALID
     cx = x + dx
     cy = y + dy
@@ -132,7 +132,7 @@ def dart_cell_neighbors(
 def dart_priorities(seed: wp.int32, out_priority: wp.array[wp.uint32]) -> None:
     # The sampling order. Drawing it up front rather than per round is what makes the whole loop a
     # deterministic function of ``seed``: every round reads the same total order on the pool.
-    i = int(wp.tid())
+    i = wp.int32(wp.tid())
     out_priority[i] = wp.randu(wp.rand_init(seed, i))
 
 
@@ -154,7 +154,7 @@ def dart_select_minima(
     # of the two" argument that keeps two acceptances apart has no exception. The pass writes only
     # its own slot, and the only state it reads of others is "not yet covered" — which this pass
     # never sets — so it is race-free and its output depends only on the round's input state.
-    t = int(wp.tid())
+    t = wp.int32(wp.tid())
     i = alive[t]
     my_key = priority[i]
     p = pool_points[i]
@@ -192,7 +192,7 @@ def dart_cover_neighbors(
     #
     # As above, only the thread's own slot is written, and a slot going ``ALIVE -> COVERED`` cannot
     # change another thread's ``== ACCEPTED`` test.
-    t = int(wp.tid())
+    t = wp.int32(wp.tid())
     i = alive[t]
     if out_state[i] != DART_ALIVE:
         return
@@ -217,7 +217,7 @@ def dart_alive_flags(
 ) -> None:
     # 0/1 survivor flags over the *work list*, in the dtype ``wp.utils.array_scan`` wants, so the
     # exclusive scan of them is directly the compaction's write positions.
-    t = int(wp.tid())
+    t = wp.int32(wp.tid())
     out_flag[t] = wp.where(state[alive[t]] == DART_ALIVE, wp.int32(1), wp.int32(0))
 
 
@@ -231,7 +231,7 @@ def dart_compact_alive(
     # Next round's work list, from this round's: survivors keep their relative order, so the loop
     # walks a shrinking prefix instead of the whole pool. ``positions`` is the exclusive scan of the
     # survivor flags.
-    t = int(wp.tid())
+    t = wp.int32(wp.tid())
     i = alive[t]
     if state[i] == DART_ALIVE:
         out_next[positions[t]] = i
